Messing with controls.
Right click now scrolls around the field, and only opnes the context menu if
you've not moved the field. Left click can be used to batch raise/kill
cells, and will remember which action you're doing.

There a block where zooming is going to go, but there's no logic for it yet.